adds anonymous closure; prevents onload overwrite
This encapsulates the entire script in an anonymous closure so that any functions declared will not overwrite (or be overwritten by) other functions from other scripts. Similarly, it now checks jQuery before adding a module, in case that module already exists. And it doesn't completely overwrite the window.onload function; rather, it adds an event listener so that other onload functions from other scripts continue to run.